requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Azure requirementes
azure-functions
# Python build requirements
wheel
# Particle filter
anglewrapper 
python-ins
